refactor(server): isolate connector hosted management
